feature : feature : Self-service invoicing rework + Purchase Order Registration Page + ShippingFeeBillableNotificationJob

+++ jeecg-module-system/jeecg-system-biz/src/main/java/org/jeecg/modules/business/controller/admin/PurchaseOrderController.java
@@ -8,20 +8,15 @@ import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.shiro.SecurityUtils;
-import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.jeecg.common.api.vo.Result;
 import org.jeecg.common.aspect.annotation.AutoLog;
 import org.jeecg.common.system.query.QueryGenerator;
 import org.jeecg.common.system.vo.LoginUser;
 import org.jeecg.common.util.oConvertUtils;
 import org.jeecg.modules.business.controller.UserException;
-import org.jeecg.modules.business.domain.purchase.invoice.InvoiceData;
 import org.jeecg.modules.business.entity.*;
 import org.jeecg.modules.business.service.*;
-import org.jeecg.modules.business.vo.InventoryImport;
-import org.jeecg.modules.business.vo.PromotionCouple;
-import org.jeecg.modules.business.vo.PurchaseOrderPage;
-import org.jeecg.modules.business.vo.SkuQuantity;
+import org.jeecg.modules.business.vo.*;
 import org.jeecg.modules.business.vo.clientPlatformOrder.PurchaseConfirmation;
 import org.jeecg.modules.business.vo.clientPlatformOrder.section.ClientInfo;
 import org.jeecgframework.poi.excel.ExcelImportUtil;
@@ -55,27 +50,22 @@ import java.util.stream.Collectors;
 @RequestMapping("/purchaseOrder")
 @Slf4j
 public class PurchaseOrderController {
-    private final IPurchaseOrderService purchaseOrderService;
-    private final IPurchaseOrderSkuService purchaseOrderSkuService;
-    private final ISkuPromotionHistoryService skuPromotionHistoryService;
-    private final ISkuService skuService;
-    private final IImportedInventoryService importedInventoryService;
-    private final IClientService clientService;
-    private final IPlatformOrderService platformOrderService;
-
     @Autowired
-    public PurchaseOrderController(IPurchaseOrderService purchaseOrderService,
-                                   IPurchaseOrderSkuService purchaseOrderSkuService,
-                                   ISkuPromotionHistoryService skuPromotionHistoryService, ISkuService skuService,
-                                   IImportedInventoryService importedInventoryService, IClientService clientService, IPlatformOrderService platformOrderService) {
-        this.purchaseOrderService = purchaseOrderService;
-        this.purchaseOrderSkuService = purchaseOrderSkuService;
-        this.skuPromotionHistoryService = skuPromotionHistoryService;
-        this.skuService = skuService;
-        this.importedInventoryService = importedInventoryService;
-        this.clientService = clientService;
-        this.platformOrderService = platformOrderService;
-    }
+    private IPurchaseOrderService purchaseOrderService;
+    @Autowired
+    private IPurchaseOrderSkuService purchaseOrderSkuService;
+    @Autowired
+    private ISkuPromotionHistoryService skuPromotionHistoryService;
+    @Autowired
+    private ISkuService skuService;
+    @Autowired
+    private IImportedInventoryService importedInventoryService;
+    @Autowired
+    private IClientService clientService;
+    @Autowired
+    private IPlatformOrderService platformOrderService;
+    @Autowired
+    private IShippingInvoiceService shippingInvoiceService;
 
     /**
      * Page query for purchase order
@@ -121,12 +111,21 @@ public class PurchaseOrderController {
     @AutoLog(value = "商品采购订单-添加")
     @ApiOperation(value="商品采购订单-添加", notes="商品采购订单-添加")
     @PostMapping(value = "/addPurchaseAndOrder")
-    public Result<?> addPurchaseAndOrder(@RequestBody PurchaseOrderPage purchaseOrderPage) {
+    public Result<?> addPurchaseAndOrder( @RequestBody PurchaseOrderPage purchaseOrderPage) {
         PurchaseOrder purchaseOrder = new PurchaseOrder();
         BeanUtils.copyProperties(purchaseOrderPage, purchaseOrder);
-        List<String> platformOrderIds = new ArrayList<>(Arrays.asList(purchaseOrderPage.getPlatformOrderId().split(",")));
+        purchaseOrder.setPaymentDocumentString(new String(purchaseOrderPage.getPaymentDocument()));
+        purchaseOrder.setInventoryDocumentString(new String(purchaseOrderPage.getInventoryDocument()));
+        String purchaseID = UUID.randomUUID().toString();
+        purchaseOrder.setId(purchaseID);
+        purchaseOrderService.save(purchaseOrder);
+        if(purchaseOrderPage.getPlatformOrderId() == null) {
+            return Result.OK("sys.api.entryAddSuccess");
+        }
+        List<String> platformOrderIds = Arrays.stream(purchaseOrderPage.getPlatformOrderId().split(","))
+                .map(String::trim)
+                .collect(Collectors.toList());
         log.info("Creating new purchase order and attributing it to orders: {}", platformOrderIds);
-        // test 5757661446491 & 5757658267995 poId
         List<PlatformOrder> platformOrders = platformOrderService.selectByPlatformOrderIds(platformOrderIds);
         Map<String, List<String>> platformOrderIdUpdateMap = new HashMap<>();
         if(!platformOrders.isEmpty()) {
@@ -145,8 +144,6 @@ public class PurchaseOrderController {
             log.error("Platform orders not found: {}", platformOrderIds);
             platformOrderIdUpdateMap.put("fail", platformOrderIds);
         }
-
-        purchaseOrderService.save(purchaseOrder);
         return Result.OK("sys.api.entryAddSuccess", platformOrderIdUpdateMap);
     }
 //    /**
@@ -187,17 +184,23 @@ public class PurchaseOrderController {
      * @param purchaseOrderPage
      * @return
      */
+    @Transactional
     @AutoLog(value = "purchase_order-编辑")
     @ApiOperation(value="purchase_order-编辑", notes="purchase_order-编辑")
     @RequestMapping(value = "/editPurchaseAndOrder", method = {RequestMethod.PUT,RequestMethod.POST})
     public Result<?> editPurchaseAndOrder(@RequestBody PurchaseOrderPage purchaseOrderPage) {
         PurchaseOrder purchaseOrder = new PurchaseOrder();
         BeanUtils.copyProperties(purchaseOrderPage, purchaseOrder);
+        purchaseOrder.setPaymentDocumentString(new String(purchaseOrderPage.getPaymentDocument()));
+        purchaseOrder.setInventoryDocumentString(new String(purchaseOrderPage.getInventoryDocument()));
+        if(purchaseOrderPage.getPlatformOrderId() == null) {
+            purchaseOrderService.updateById(purchaseOrder);
+            return Result.OK("sys.api.entryEditSuccess");
+        }
         List<String> platformOrderIds = new ArrayList<>(Arrays.asList(purchaseOrderPage.getPlatformOrderId().split(",")));
         log.info("Editing purchase order and attributing it to orders : {}", platformOrderIds);
         log.info("Removing previous attribution to orders");
         platformOrderService.removePurchaseInvoiceNumber(purchaseOrder.getInvoiceNumber());
-        // test 5757661446491 & 5757658267995 poId
         List<PlatformOrder> platformOrders = platformOrderService.selectByPlatformOrderIds(platformOrderIds);
         log.info("Platform orders found for attribution : {}", platformOrders.stream().map(PlatformOrder::getPlatformOrderId).collect(Collectors.toList()));
         Map<String, List<String>> platformOrderIdUpdateMap = new HashMap<>();
@@ -216,7 +219,6 @@ public class PurchaseOrderController {
             log.error("Platform orders not found: {}", platformOrderIds);
             platformOrderIdUpdateMap.put("fail", platformOrderIds);
         }
-
         purchaseOrderService.updateById(purchaseOrder);
         return Result.OK("sys.api.entryEditSuccess", platformOrderIdUpdateMap);
     }
@@ -231,6 +233,11 @@ public class PurchaseOrderController {
     @ApiOperation(value = "商品采购订单-通过id删除", notes = "商品采购订单-通过id删除")
     @DeleteMapping(value = "/delete")
     public Result<?> delete(@RequestParam(name = "id", required = true) String id) {
+        PurchaseOrder po = purchaseOrderService.getById(id);
+        if(po.getInventoryDocumentString() != null && !po.getInventoryDocumentString().isEmpty())
+            shippingInvoiceService.deleteAttachmentFile(po.getInventoryDocumentString());
+        if(po.getPaymentDocumentString() != null && !po.getPaymentDocumentString().isEmpty())
+            shippingInvoiceService.deleteAttachmentFile(po.getPaymentDocumentString());
         purchaseOrderService.delMain(id);
         return Result.OK("sys.api.entryDeleteSuccess");
     }
@@ -245,8 +252,54 @@ public class PurchaseOrderController {
     @ApiOperation(value = "商品采购订单-批量删除", notes = "商品采购订单-批量删除")
     @DeleteMapping(value = "/deleteBatch")
     public Result<?> deleteBatch(@RequestParam(name = "ids", required = true) String ids) {
+        List<PurchaseOrder> purchaseOrders = purchaseOrderService.listByIds(Arrays.asList(ids.split(",")));
+        for(PurchaseOrder po : purchaseOrders) {
+            if(po.getInventoryDocumentString() != null && !po.getInventoryDocumentString().isEmpty())
+                shippingInvoiceService.deleteAttachmentFile(po.getInventoryDocumentString());
+            if(po.getPaymentDocumentString() != null && !po.getPaymentDocumentString().isEmpty())
+                shippingInvoiceService.deleteAttachmentFile(po.getPaymentDocumentString());
+        }
         this.purchaseOrderService.delBatchMain(Arrays.asList(ids.split(",")));
         return Result.OK("sys.api.entryBatchDeleteSuccess");
+    }
+
+    /**
+     * Cancel a purchase order's invoice.
+     * @param purchaseId purchase ID
+     * @param invoiceNumber invoice number
+     * @return
+     */
+    @AutoLog(value = "商品采购订单-通过id删除")
+    @ApiOperation(value = "商品采购订单-通过id删除", notes = "商品采购订单-通过id删除")
+    @DeleteMapping(value = "/cancelInvoice")
+    public Result<?> cancelInvoice(@RequestParam("id") String purchaseId,
+                                   @RequestParam("invoiceNumber") String invoiceNumber) {
+        PurchaseOrder po = purchaseOrderService.getById(purchaseId);
+        if(po.getInventoryDocumentString() != null && !po.getInventoryDocumentString().isEmpty())
+            shippingInvoiceService.deleteAttachmentFile(po.getInventoryDocumentString());
+        if(po.getPaymentDocumentString() != null && !po.getPaymentDocumentString().isEmpty())
+            shippingInvoiceService.deleteAttachmentFile(po.getPaymentDocumentString());
+        purchaseOrderService.cancelInvoice(purchaseId, invoiceNumber);
+        return Result.OK("sys.api.entryDeleteSuccess");
+    }
+    /**
+     * Cancel multiple purchase order's invoice.
+     * @param purchaseIds list of purchase IDs
+     * @return
+     */
+    @AutoLog(value = "商品采购订单-批量删除")
+    @ApiOperation(value = "商品采购订单-批量删除", notes = "商品采购订单-批量删除")
+    @DeleteMapping(value = "/cancelBatchInvoice")
+    public Result<?> cancelBatchInvoice(@RequestParam("ids") String purchaseIds) {
+        List<PurchaseOrder> purchaseOrders = purchaseOrderService.listByIds(Arrays.asList(purchaseIds.split(",")));
+        for(PurchaseOrder po : purchaseOrders) {
+            if(po.getInventoryDocumentString() != null && !po.getInventoryDocumentString().isEmpty())
+                shippingInvoiceService.deleteAttachmentFile(po.getInventoryDocumentString());
+            if(po.getPaymentDocumentString() != null && !po.getPaymentDocumentString().isEmpty())
+                shippingInvoiceService.deleteAttachmentFile(po.getPaymentDocumentString());
+        }
+        purchaseOrderService.cancelBatchInvoice(purchaseIds);
+        return Result.OK("sys.api.entryDeleteSuccess");
     }
 
     /**
@@ -447,7 +500,7 @@ public class PurchaseOrderController {
      * @param purchaseID purchaseID
      */
     @RequestMapping(value = "/invoiceMeta", method = RequestMethod.GET)
-    public InvoiceData getInvoiceMetaData(@RequestParam String purchaseID, HttpServletResponse response) throws IOException, URISyntaxException {
+    public InvoiceMetaData getInvoiceMetaData(@RequestParam String purchaseID, HttpServletResponse response) throws IOException, URISyntaxException {
         return purchaseOrderService.makeInvoice(purchaseID);
     }
